Working on residuals (end of day commit)
Using residual: diag(conj(v))*Y*v - conj(S)
to evlaulate the linearization.
TODO
-----------
there appear to be mainly errors in the reactive power injection
for PV buses, that is: scatter(Qgreal, Qglin) is crazy.
However, if simply the voltage magnitudes are replaced the solution
seems to improve dramatically. This is particularly intersting given
that this appears to be the case even when the voltage magnitude
solution is quite good (e.g polish 3375wp with correlation of 0.9911).
Improving Qinj is probably where efforts should go.